requirements.txt
numpy
Pillow
trimesh